requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
flask==2.3.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
newsapi-python==0.2.7
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
werkzeug==2.3.7
zipp==3.16.2
